% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{commit_activity_df}
\alias{commit_activity_df}
\title{Get all contributor commit activity}
\usage{
commit_activity_df(activity, contributors = NULL, collapse_to = "week")
}
\arguments{
\item{activity}{\code{list} from \code{get_all_contributor_commit_activity}}

\item{contributors}{\code{chr} of contributors to include}

\item{collapse_to}{\code{chr} one of \code{week}, \code{authors}, \code{repos} where \code{week} collapses the week field to a data.frame. \code{authors} collapses the week field and authors field to a data.frame with \code{authors} as a column. \code{repos} collapses the week field, authors field, and repos to a data.frame with \code{authors} and \code{week} as columns}
}
\value{
\code{list} of all activity
}
\description{
Get all contributor commit activity
}
