% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commits.R
\name{commit_get}
\alias{commit_get}
\title{Title}
\usage{
commit_get(owner, repo, ref, page = 1, per_page = 100)
}
\arguments{
\item{owner}{\code{chr} The account owner of the repository. The name is not case sensitive.}

\item{repo}{\code{chr} The name of the repository. The name is not case sensitive.}

\item{ref}{\code{chr} Sha of the commit}

\item{page}{\code{num} Page number of the results to fetch.}

\item{per_page}{\code{num} The number of results per page (max 100).}
}
\value{
\code{list} of commit details
}
\description{
Title
}
