% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commits.R
\name{commits_list}
\alias{commits_list}
\title{Get Commit List}
\usage{
commits_list(
  owner,
  repo,
  author = NULL,
  sha = NULL,
  path = NULL,
  since = NULL,
  until = NULL,
  per_page = 100,
  page = 1
)
}
\arguments{
\item{owner}{\code{chr} The account owner of the repository. The name is not case sensitive.}

\item{repo}{\code{chr} The name of the repository. The name is not case sensitive.}

\item{author}{\code{chr} GitHub login or email address by which to filter by commit author.}

\item{sha}{\code{chr} SHA or branch to start listing commits from. Default: the repository’s default branch (usually master).}

\item{path}{\code{chr} Only commits containing this file path will be returned.}

\item{since}{\code{chr/POSIXct} Only show notifications updated after the given time. This is a timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ.}

\item{until}{\code{chr/POXIXct} Only commits before this date will be returned. This is a timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ.}

\item{per_page}{\code{num} The number of results per page (max 100).}

\item{page}{\code{num} Page number of the results to fetch.}
}
\value{
\code{list}
}
\description{
See \href{List Commits}{https://docs.github.com/en/rest/commits/commits#list-commits}
}
