% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{get_branches}
\alias{get_branches}
\title{Get branch info}
\usage{
get_branches(
  owner,
  repo,
  protected = NULL,
  per_page = 100,
  page = 1,
  branch = NULL,
  branches = NULL
)
}
\arguments{
\item{owner}{\code{chr} The account owner of the repository. The name is not case sensitive.}

\item{repo}{\code{chr} The name of the repository. The name is not case sensitive.}

\item{protected}{\code{lgl} Setting to \code{true} returns only protected branches. When set to \code{false}, only unprotected branches are returned. Omitting this parameter returns all branches.}

\item{per_page}{\code{num} The number of results per page (max 100).}

\item{page}{\code{num} Page number of the results to fetch.}

\item{branch}{\code{chr} Branch SHA, for querying info about a single branch}

\item{branches}{\code{list} of Branches, to concatenate in a paginated query. Mostly of internal use in other functions.}
}
\value{
\code{list} of Branch info (does not include file info)
}
\description{
Get branch info
}
\seealso{
\href{https://docs.github.com/en/rest/branches/branches}{Github API Documentation}
}
