% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{get_list_all_organization_repositories}
\alias{get_list_all_organization_repositories}
\title{List Organization repositories}
\usage{
get_list_all_organization_repositories(
  org,
  type = NULL,
  sort = NULL,
  direction = NULL,
  per_page = 100,
  page = NULL
)
}
\arguments{
\item{type}{\code{chr} Specifies the types of repositories you want returned. If your organization is associated with an enterprise account using GitHub Enterprise Cloud or GitHub Enterprise Server 2.20+, type can also be internal. However, the internal value is not yet supported when a GitHub App calls this API with an installation access token.
Can be one of: \code{all}, \code{public}, \code{private}, \code{forks}, \code{sources}, \code{member}, \code{internal}}

\item{sort}{\code{chr} What to sort results by. \code{popularity} will sort by the number of comments. \code{long-running} will sort by date created and will limit the results to pull requests that have been open for more than a month and have had activity within the past month.
Default: \code{created}
Can be one of: \verb{created, updated, popularity, long-running}}

\item{direction}{\code{chr} The direction of the sort. Default: \code{desc} when sort is created or sort is not specified, otherwise \code{asc}.
Can be one of: \code{asc}, \code{desc}}

\item{per_page}{\code{num} The number of results per page (max 100).}

\item{page}{\code{num} Page number of the results to fetch.}
}
\value{
\code{list} of repositories
}
\description{
List Organization repositories
}
