% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulls.R
\name{get_pulls}
\alias{get_pulls}
\title{Get PRs}
\usage{
get_pulls(
  owner,
  repo,
  state = "open",
  head = NULL,
  base = NULL,
  sort = "created",
  direction = "desc",
  per_page = 100,
  page = NULL
)
}
\arguments{
\item{owner}{\code{chr} The account owner of the repository. The name is not case sensitive.}

\item{repo}{\code{chr} The name of the repository. The name is not case sensitive.}

\item{state}{\code{chr} Either \code{open}, \code{closed}, or \code{all} to filter by state.
Default: \code{open}
Can be one of: \verb{open, closed, all}}

\item{head}{\code{chr} Filter pulls by head user or head organization and branch name in the format of \code{user:ref-name} or \code{organization:ref-name}. For example: \code{github:new-script-format} or \code{octocat:test-branch}.}

\item{base}{\code{chr} Filter pulls by base branch name. Example: \code{gh-pages}.}

\item{sort}{\code{chr} What to sort results by. \code{popularity} will sort by the number of comments. \code{long-running} will sort by date created and will limit the results to pull requests that have been open for more than a month and have had activity within the past month.
Default: \code{created}
Can be one of: \verb{created, updated, popularity, long-running}}

\item{direction}{\code{chr} The direction of the sort. Default: \code{desc} when sort is created or sort is not specified, otherwise \code{asc}.
Can be one of: \code{asc}, \code{desc}}

\item{per_page}{\code{num} The number of results per page (max 100).}

\item{page}{\code{num} Page number of the results to fetch.}
}
\value{
\code{list} of pulls
}
\description{
Get PRs
}
\seealso{
\href{https://docs.github.com/en/rest/pulls/pulls}{Github API Documentation}
}
