% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{total_contributions_summary}
\alias{total_contributions_summary}
\title{Retrieve a summary of contributions for a repo}
\usage{
total_contributions_summary(
  owner,
  repo,
  author = NULL,
  since = NULL,
  until = NULL,
  file_types = c("R", "Rmd", "js", "scss", "svelte"),
  exclude_main = TRUE,
  enable_resume = TRUE,
  include_all_unique_commit_details = FALSE,
  include_branch_details = FALSE,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{owner}{\code{chr} The account owner of the repository. The name is not case sensitive.}

\item{repo}{\code{chr} The name of the repository. The name is not case sensitive.}

\item{author}{\code{chr} GitHub login or email address by which to filter by commit author.}

\item{since}{\code{chr/POSIXct} Only show notifications updated after the given time. This is a timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ.}

\item{until}{\code{chr/POXIXct} Only commits before this date will be returned. This is a timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ.}

\item{file_types}{\code{chr} file types to consider}

\item{exclude_main}{\code{lgl} Whether to filter out the main branch. Default \code{TRUE}}

\item{enable_resume}{\code{lgl} Whether to fetch API calls from the mock API repository if they've been saved in a previous run.}

\item{include_all_unique_commit_details}{\code{lgl} Whether to include the details about all unique commits contributing to the summary.}

\item{include_branch_details}{\code{lgl} Whether to include the commit list for each branch.}

\item{...}{reserved for future arguments.}

\item{verbose}{\code{lgl} whether to print status messages}
}
\value{
\code{list}
}
\description{
Retrieve a summary of contributions for a repo
}
