% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weekly.R
\name{weekly_prs}
\alias{weekly_prs}
\title{Generate a Markdown list of all PRs and their requestors since Date}
\usage{
weekly_prs(
  owner = "Martin-McCoy",
  repo = "dmdu",
  state = "all",
  since = lubridate::floor_date(lubridate::as_date(lubridate::today() -
    lubridate::weeks(1), regex_remove = "[dD]ev to [mM]ain"), "week")
)
}
\arguments{
\item{owner}{\code{chr} The account owner of the repository. The name is not case sensitive.}

\item{repo}{\code{chr} The name of the repository. The name is not case sensitive.}

\item{state}{\code{chr} Either \code{open}, \code{closed}, or \code{all} to filter by state.
Default: \code{open}
Can be one of: \verb{open, closed, all}}

\item{since}{\code{Date} Return all PRs since this Date}
}
\value{
\code{chr}
}
\description{
Generate a Markdown list of all PRs and their requestors since Date
}
\examples{
weekly_prs()
}
